requirements.txt
cryptography == 2.7
paho-mqtt == 1.4.0
PyBluez == 0.22
RPi.GPIO == 0.7.0
mock==3.0.5
numexpr==2.7.0
numpy==1.17.2
six==1.12.0
tables==3.5.2